+++ src/modules/manufacturing/entities/kp-oem-mesin-sortir.ts
@@ -1,0 +1,34 @@
+import { Field, ObjectType } from 'type-graphql';
+import { BaseEntity, Column, Entity, PrimaryColumn } from 'typeorm';
+
+@Entity('VKY_KP_OEM_MESIN_SORTIR')
+@ObjectType()
+export class KpOemMesinSortir extends BaseEntity {
+  @Field()
+  @PrimaryColumn({ name: 'CONTRACT' })
+  contract!: string;
+
+  @Field()
+  @PrimaryColumn({ name: 'ORDER_NO' })
+  orderNo!: string;
+
+  @Field()
+  @PrimaryColumn({ name: 'LOT_BATCH_NO' })
+  lotBatchNo!: string;
+
+  @Field()
+  @Column({ name: 'QUANTITY' })
+  quantity!: number;
+
+  @Field()
+  @Column({ name: 'STATUS' })
+  status!: string;
+
+  @Field()
+  @Column({ name: 'CREATED_DATE' })
+  createdDate!: Date;
+
+  @Field()
+  @Column({ name: 'MODIFIED_DATE' })
+  modifiedDate!: Date;
+}